Refactor instruction retirement.
The main goal here is to surface wait-states of the hart
(e.g. as caused by WFI and WRS.{NTO,STO}) into the stepper.

The stepper is the main interface to the Sail-external harness.
This makes it easier to communicate wait-states to the harness,
allowing the pausing of Sail execution could be made configurable.

The refactor makes interrupt and exception handling more
symmetrical: both are now handled in the stepper.

The type related to retirement failure is now in its own file.

This also removes the model-internal global instbits register into
a stepper-local variable.