Add a cosimulator API for lockstep simulation with another simulator.
This builds on the callbacks from the model to drive the cosimulator.

Implement the cosimulator API for the Spike simulator.  This implements
GPR and CSR cross-checks.  Floating-point and vector checks are left
for later.

Add a USE_SPIKE build option for the Spike cosimulator; when disabled
(default), use a null cosimulator.  When built to link against the Spike
libraries, the Spike cosimulator is enabled with an `--enable-spike`
cli option.

The Spike install is located using the $RISCV environment variable
described in the Spike build instructions.  This needs to be in
the environment for the USE_SPIKE build option.